+++ packages/dso-toolkit/components/Componenten/tooltip/tooltip.config.js
@@ -3,13 +3,6 @@ module.exports = {
   meta: {
     webComponent: 'dso-tooltip'
   },
-  notes: `
-Een tooltip hoort bij een \`target\` element. Dit kan aangegeven worden met het attribuut \`for\`. Dit kan een \`string\` zijn met een id of een \`HTMLElement\`. Zonder \`target\` wordt de parent van \`<dso-tooltip>\` gebruikt.
-
-Scriptend de tooltip tonen/verbergen kan met de instance methods \`activate()\` en \`deactivate()\` of met het het attribuut \`active\`.
-
-Het positioneren van de tooltip wordt met [Popper](https://popper.js.org/) gedaan.
-  `,
   collated: true,
   collator: (markup, item) => `<div style="position: relative; height: 4em;">${markup}</div>`,
   context: {